Implemented most of the code
I have set up the sockets, menu, and the program can send data in one
direction so far, from the server to the client.  It utilizes a thread
for receiving data so it should not block the client from sending data
when I get around to adding a thread for that.

I also still need to add some commands like quit, save log, etc, and
make it output the time next to each message.  Maybe I'll add in
nicknames too.